+++ src/app/components/header/header.component.ts
@@ -11,6 +11,6 @@ export class HeaderComponent {
     private cartService: CartService ) {}
  
   getItemsCount() {
-    return this.cartService.ItemsCount;
+    return this.cartService.TotalCount;
   }
 }
